pyproject.toml
[tool.poetry]
name = "promptflow-starter-template"
version = "0.1.2"
description = ""
authors = ["Fabian Schurig <fabian.a.schurig@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
promptflow = "^1.17.2"
promptflow-azure = "^1.17.2"
promptflow-core = "^1.17.2"
promptflow-devkit = "^1.17.2"
promptflow-tools = "^1.5.0"
promptflow-tracing = "^1.17.2"
python-dotenv = "^1.0.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
keyrings-alt = "^5.0.2"

[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"
pylint = "^3.3.4"
pytest = "^8.3.4"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
pytest-mypy = "^0.10.3"
setuptools = "^75.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
